添加分类，更新分类功能

+++ pocket-shop-web-admin/src/main/java/cn/ishangit/shop/web/admin/web/controller/TbContentController.java
@@ -29,7 +29,7 @@ public class TbContentController {
     private TbContentService tbContentService;
 
     @ModelAttribute
-    public TbContent getTbContent(Integer id){
+    public TbContent getTbContent(Long id){
         TbContent tbContent = new TbContent();
         if (id != null){
             tbContent = tbContentService.getById(id);
